use new Template class
