+++ pyftp/modeler.py
@@ -30,227 +30,6 @@ def LMfit(x, y, err, cn, sn, omega, sgn=1):
     return a, b, c
 
 
-def rms_resid_over_rms(cn, sn, Tt, Yt):
-    # This is fairly slow; is there a better way to get best fit pars?
-    a, b, c = LMfit(Tt, Yt, np.ones(len(Tt))*0.0001, cn, sn, 2 * np.pi, sgn=True)
-    Ym = ftp.fitfunc(Tt, 1, 2 * np.pi, cn, sn, a, b, c)
-
-    S = sqrt(np.mean(np.power(Yt, 2)))
-
-    Rp = sqrt(np.mean(np.power(Ym - Yt, 2))) / S
-
-    a, b, c = LMfit(Tt, Yt, np.ones(len(Tt))*0.0001, cn, sn, 2 * np.pi, sgn=False)
-    Ym = ftp.fitfunc(Tt, -1, 2 * np.pi, cn, sn, a, b, c)
-
-    Rn = sqrt(np.mean(np.power(Ym - Yt, 2))) / S
-    return min([ Rn, Rp ])
-
-rms = lambda x : sqrt(np.mean(np.power(x, 2)))
-
-
-def match_up_truncated_template(cn, sn, Tt, Yt):
-    Ym = ftp.fitfunc(Tt, 1, 2 * np.pi, cn, sn, 2.0, 0.0, 0.0)
-
-    # Align the maxima of truncated and full templates
-    di = np.argmax(Ym) - np.argmax(Yt)
-
-    # Add some 'wiggle room', since maxima may be offset by 1
-    Ym = [ np.array([ Ym[(j + (di + k))%len(Ym)] for j in range(len(Ym)) ]) for k in [ -1, 0, 1 ] ]
-
-    # Align the heights of the truncated and full templates
-    Ym = [ Y + (Yt[0] - Y[0]) for Y in Ym ]
-
-    # Keep the best fit
-    return Ym[np.argmin( [ rms(Y - Yt) for Y in Ym ] )]
-
-
-def rms_resid_over_rms_fast(cn, sn, Tt, Yt):
-    Ym = match_up_truncated_template(cn, sn, Tt, Yt)
-    return rms(Yt - Ym) / rms(Yt)
-
-
-def approximate_template(Tt, Yt, errfunc=rms_resid_over_rms, stop=1E-2, nharmonics=None):
-    """ Fourier transforms template, returning the first H components """
-
-    #print "fft"
-    fft = np.fft.rfft(Yt)
-
-    cn, sn = None, None
-    if not nharmonics is None and int(nharmonics) > 0:
-        #print "creating cn and sn"
-        cn, sn = zip(*[ (p.real/len(Tt), -p.imag/len(Tt)) for i,p in enumerate(fft) \
-                     if i > 0 and i <= int(nharmonics) ])
-
-    else:
-
-        cn, sn = zip(*[ (p.real/len(Tt), -p.imag/len(Tt)) for i,p in enumerate(fft) \
-                     if i > 0 ])
-
-        h = 1
-        while errfunc(cn[:h], sn[:h], Tt, Yt) > stop:
-            #print "h -> ", h
-            h+=1
-
-        cn, sn = cn[:h], sn[:h]
-    return np.array(cn), np.array(sn)
-
-normfac = lambda cn, sn : 1./np.sqrt(sum([ ss*ss + cc*cc for cc, ss in zip(cn, sn) ]))
-
-
-class Template(object):
-    """
-    Template class
-
-    y(t) = sum[n]( c[n]cos(nwt) + s[n]sin(nwt) )
-
-    Parameters
-    ----------
-    cn: array-like, optional
-        Truncated Fourier coefficients (cosine)
-    sn: array-like, optional
-        Truncated Fourier coefficients (sine)
-    phase: array-like, optional
-        phase-values, must contain floating point numbers in [0,1]
-    y: array-like, optional
-        amplitude of template at each phase value
-    stop: float, optional (default: 2E-2)
-        will pick minimum number of harmonics such that
-        rms(trunc(template) - template) / rms(template) < stop
-    nharmonics: None or int, optional (default: None)
-        Keep a constant number of harmonics
-    fname: str, optional
-        Filename to load/save template
-    errfunc: callable, optional (default: rms_resid_over_rms)
-        A function returning some measure of error resulting
-        from approximating the template with a given number
-        of harmonics
-    template_id: str, optional
-        Name of template
-
-    """
-    def __init__(self, cn=None, sn=None, phase=None, y=None,
-                       stop=2E-2, nharmonics=None, fname=None,
-                       errfunc=rms_resid_over_rms, template_id=None):
-
-        self.phase = phase
-        self.y = y
-
-        self.fname = fname
-        self.stop = stop
-        self.nharmonics = nharmonics
-        self.errfunc = errfunc
-        self.cn = None
-        self.sn = None
-        self.pvectors = None
-        self.ptensors = None
-        self.template_id = template_id
-        self.best_fit_y = None
-
-        self.cn_full = None
-        self.sn_full = None
-
-
-        self.defined_by_ty = cn is None and sn is None
-        if self.defined_by_ty and (phase is None or y is None):
-            raise Exception("Need to define either (phase, y) or (cn, sn) for template")
-
-        if not self.defined_by_ty:
-
-            assert(not (cn is None or sn is None))
-
-            self.cn_full = np.array(cn).copy()
-            self.sn_full = np.array(sn).copy()
-
-            if nharmonics is None:
-                self.nharmonics = len(cn)
-
-            assert(self.nharmonics <= len(cn))
-
-            self.cn = self.cn_full[:self.nharmonics].copy()
-            self.sn = self.sn_full[:self.nharmonics].copy()
-
-
-            self.cn *= normfac(self.cn, self.sn)
-            self.sn *= normfac(self.cn, self.sn)
-
-    def is_saved(self):
-
-        return (not self.fname is None and os.path.exists(self.fname))
-
-    def precompute(self):
-
-
-        if self.defined_by_ty:
-            #print "approximating template"
-            self.cn, self.sn = approximate_template(self.phase, self.y,
-                                        stop=self.stop, errfunc=self.errfunc,
-                                        nharmonics=self.nharmonics)
-
-            #print "getting best fit y"
-            self.best_fit_y = match_up_truncated_template(self.cn, self.sn, self.phase, self.y)
-
-            #print "computing rms_resid/rms"
-            self.rms_resid_over_rms = rms(self.best_fit_y - self.y) / rms(self.y)
-
-        else:
-
-            assert(self.nharmonics <= len(self.cn_full))
-
-            self.cn = self.cn_full[:self.nharmonics].copy()
-            self.sn = self.sn_full[:self.nharmonics].copy()
-
-            self.cn *= normfac(self.cn, self.sn)
-            self.sn *= normfac(self.cn, self.sn)
-
-            nph = 100
-            self.phase = np.linspace(0, 1, nph)
-            self.y = ftp.fitfunc(self.phase, 1, 2 * np.pi, self.cn_full, self.sn_full, 2.0, 0.0, 1.0)
-            self.best_fit_y = ftp.fitfunc(self.phase, 1, 2 * np.pi, self.cn, self.sn, 2.0, 0.0, 1.0)
-
-            self.rms_resid_over_rms = rms(self.best_fit_y - self.y) / rms(self.y)
-
-        #print "computing pvectors"
-        #print self.cn
-        #print self.sn
-        self.pvectors = ftp.get_polynomial_vectors(self.cn, self.sn, sgn=1)
-
-        #print "computing ptensors"
-        self.ptensors = ftp.compute_polynomial_tensors(*self.pvectors)
-
-        return self
-
-    def load(self, fname=None):
-        fn = fname if not fname is None else self.fname
-        self.__dict__.update(pickle.load(open(fn, 'rb')))
-
-    def save(self, fname=None):
-        fn = fname if not fname is None else self.fname
-        pickle.dump(self.__dict__, open(fn, 'wb'))
-
-    def add_plot_to_axis(self, ax):
-        ax.plot(self.phase, self.y, color='k', label="original")
-        ax.plot(self.phase, self.best_fit_y,
-                  label="truncated (H=%d)"%(self.nharmonics))
-
-    def set_nharmonics(self, nharmonics):
-        self.nharmonics = nharmonics
-        self.precompute()
-        return self
-
-    def plot(self, plt):
-        f, ax = plt.subplots()
-        self.add_plot_to_axis(ax)
-        ax.set_xlim(0, 1)
-        #ax.set_ylim(0, 1)
-        ax.set_xlabel('phase')
-        ax.set_ylabel('y')
-        ax.set_title('"%s", stop = %.3e, H = %d'%(self.template_id,
-                                             self.stop, self.nharmonics))
-        ax.legend(loc='best', fontsize=9)
-        plt.show()
-        plt.close(f)
-
-
 class FastTemplateModeler(object):
 
     """
